+++ src/parser/properties/attendee.parser.ts
@@ -12,8 +12,9 @@ const validParameters: Parameters = {
 	'ROLE': 'role',
 	'PARTSTAT': 'partstat',
 	'RSVP': 'rsvp',
-	'DELEGATED-TO': 'delegatedTo',
-	'DELEGATED-FROM': 'delegatedFrom',
+	'DELEGATED-TO': { name: 'delegatedTo', lambda: (text: string): string => text.replace('mailto:', '') },
+	'DELEGATED-FROM': { name: 'delegatedFrom', lambda: (text: string): string => text.replace('mailto:', '') },
+	'SCHEDULE-STATUS': { name: 'scheduleStatus', lambda: (text: string): string | string[] => text.includes(',') ? text.split(',') : text }
 }
 
 class AttendeeParser extends BaseParser<Attendee> {
@@ -21,6 +22,8 @@ class AttendeeParser extends BaseParser<Attendee> {
 		if (!value) {
 			throw new ParsingError('Empty iCalendar attendee value')
 		}
+
+		value = value.replace('mailto:', '')
 
 		return {
 			address: value,